Rename interpreter
